Rasberrypicar
This was made during hackpsu 2019 spring.  The goal of this was to get a rasberry pi car to drive around with google home mini commands.  I was not able to figure out how to get the google home mini to interact with firebase I was not able to code that part.